afyalink_backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
gunicorn
dj-database-url
psycopg2-binary
whitenoise
